customizado

+++ frontend/lista-de-tarefas/src/components/pages/CreatList/CreatList.jsx
@@ -10,6 +10,7 @@ export function CreatList() {
   const [status, setStatus] = useState("");
   const [editingTask, setEditingTask] = useState(null);
   const [completedTasks, setCompletedTasks] = useState(new Set());
+  const [messageAlert, setMessageAlert] = useState("");
 
   const getTasks = async () => {
     try {
@@ -33,15 +34,29 @@ export function CreatList() {
     }
 
     try {
+      
       if (editingTask) {
+
+        const confirmEdit = window.confirm("Deseja realmente editar essa tarefa?")
+        if(!confirmEdit) return
+        
         await taskService.updateTask(editingTask.id, {
           titulo: title,
           categoria: category,
           status,
         });
+        setMessageAlert("Tarefa Atualizada com sucesso!✅");
       } else {
-        await taskService.createTask({ titulo: title, categoria: category, status });
+        await taskService.createTask({
+          titulo: title,
+          categoria: category,
+          status,
+        });
+        setMessageAlert("Tarefa criada com sucesso! 🎉");
       }
+      setTimeout(() => {
+        setMessageAlert("");
+      }, 3000);
       setCategory("");
       setTitle("");
       setStatus("");
@@ -54,8 +69,16 @@ export function CreatList() {
 
   const deleteTask = async (id) => {
     try {
+      const confirmDelete = window.confirm("Deletar tarefa?");
+      if (!confirmDelete) return;
+
       await taskService.deleteTask(id);
       setTaskList(taskList.filter((task) => task.id !== id));
+
+      setMessageAlert("Tarefa deletada com sucesso!✅");
+      setTimeout(() => {
+        setMessageAlert("");
+      }, 3000);
       getTasks();
     } catch (error) {
       console.error("Error deleting task", error);
@@ -75,6 +98,10 @@ export function CreatList() {
         updated.delete(id);
       } else {
         updated.add(id);
+        setMessageAlert("Tarefa completa ✅");
+        setTimeout(() => {
+          setMessageAlert("");
+        }, 3000);
       }
       return updated;
     });
@@ -85,103 +112,108 @@ export function CreatList() {
   });
 
   return (
-    <div className={style.wrapper}>
-      <div className={style.container}>
-        <div className={style.containerButtons}>
-          <div className={style.form}>
-            <form onSubmit={createOrUpdateTask}>
-              <h1 className={style.heading}>Lista de tarefas</h1>
-              <div className={style.createTask}>
-                <h4 className={style.subheading}>
-                  {editingTask ? "Editar tarefa" : "Criar tarefa"}
-                </h4>
-                <div className={style.selectCategory}>
-                  <select
-                    className={style.categoryButton}
-                    name="create"
-                    id="create"
-                    value={category}
-                    onChange={(e) => setCategory(e.target.value)}
-                  >
-                    <option value="">Selecione uma categoria</option>
-                    <option value="Estudo">Estudo</option>
-                    <option value="Trabalho">Trabalho</option>
-                    <option value="Pessoal">Pessoal</option>
-                  </select>
+    <>
+      {messageAlert && <p className={style.messagemAlert}>{messageAlert}</p>}
+      <div className={style.wrapper}>
+        <div className={style.container}>
+          <div className={style.containerButtons}>
+            <div className={style.form}>
+              <form onSubmit={createOrUpdateTask}>
+                <h1 className={style.heading}>Lista de tarefas</h1>
+                <div className={style.createTask}>
+                  <h4 className={style.subheading}>
+                    {editingTask ? "Editar tarefa" : "Criar tarefa"}
+                  </h4>
+                  <div className={style.selectCategory}>
+                    <select
+                      className={style.categoryButton}
+                      name="create"
+                      id="create"
+                      value={category}
+                      onChange={(e) => setCategory(e.target.value)}
+                    >
+                      <option value="">Selecione uma categoria</option>
+                      <option value="Estudo">Estudo</option>
+                      <option value="Trabalho">Trabalho</option>
+                      <option value="Pessoal">Pessoal</option>
+                    </select>
+                  </div>
+                  <div className={style.inputArea}>
+                    <input
+                      placeholder="Digite o titulo"
+                      type="text"
+                      id="title"
+                      value={title}
+                      onChange={(e) => setTitle(e.target.value)}
+                      required
+                    />
+                    <button type="submit" className={style.createOrEdit}>
+                      {editingTask ? "Salvar tarefa" : "Criar tarefa"}
+                    </button>
+                  </div>
                 </div>
-                <div className={style.inputArea}>
-                  <input
-                    placeholder="Digite o titulo"
-                    type="text"
-                    id="title"
-                    value={title}
-                    onChange={(e) => setTitle(e.target.value)}
-                    required
-                  />
-                  <button type="submit" className={style.createOrEdit}>
-                    {editingTask ? "Salvar tarefa" : "Criar tarefa"}
-                  </button>
-                </div>
-              </div>
-            </form>
-          </div>
-
-          <div className={style.taskContainer}>
-            <div className={style.searchTasks}>
-              <h4 className={style.searchTitle}>Pesquisar</h4>
-              <select
-                id="search"
-                value={filterCategory}
-                onChange={(e) => setFilterCategory(e.target.value)}
-              >
-                <option value="Todas">Todas</option>
-                <option value="Estudo">Estudo</option>
-                <option value="Trabalho">Trabalho</option>
-                <option value="Pessoal">Pessoal</option>
-              </select>
+              </form>
             </div>
 
-            <div className={style.taskList}>
-              {filteredTasks.length === 0 ? (
-                <p>Não ha tarefas registradas!</p>
-              ) : (
-                filteredTasks.map((task) => (
-                  <div key={task.id} className={style.taskItem}>
-                    <div
-                      className={`${style.category} ${
-                        completedTasks.has(task.id) ? style.completed : ""
-                      }`}
-                    >
-                      <p className={style.title}>{task.titulo}</p>
-                      <p className={style.categoryLabel}>({task.categoria})</p>
+            <div className={style.taskContainer}>
+              <div className={style.searchTasks}>
+                <h4 className={style.searchTitle}>Pesquisar</h4>
+                <select
+                  id="search"
+                  value={filterCategory}
+                  onChange={(e) => setFilterCategory(e.target.value)}
+                >
+                  <option value="Todas">Todas</option>
+                  <option value="Estudo">Estudo</option>
+                  <option value="Trabalho">Trabalho</option>
+                  <option value="Pessoal">Pessoal</option>
+                </select>
+              </div>
+
+              <div className={style.taskList}>
+                {filteredTasks.length === 0 ? (
+                  <p>Não ha tarefas registradas!</p>
+                ) : (
+                  filteredTasks.map((task) => (
+                    <div key={task.id} className={style.taskItem}>
+                      <div
+                        className={`${style.category} ${
+                          completedTasks.has(task.id) ? style.completed : ""
+                        }`}
+                      >
+                        <p className={style.title}>{task.titulo}</p>
+                        <p className={style.categoryLabel}>
+                          ({task.categoria})
+                        </p>
+                      </div>
+                      <div className={style.buttons}>
+                        <button
+                          className={style.complete}
+                          onClick={() => toggleCompleteTask(task.id)}
+                        >
+                          Complete
+                        </button>
+                        <button
+                          className={style.delete}
+                          onClick={() => deleteTask(task.id)}
+                        >
+                          <i className="fa-solid fa-trash"></i>
+                        </button>
+                        <button
+                          className={style.edit}
+                          onClick={() => handleEditForm(task)}
+                        >
+                          <i className="fa-solid fa-pen-to-square"></i>
+                        </button>
+                      </div>
                     </div>
-                    <div className={style.buttons}>
-                      <button
-                        className={style.complete}
-                        onClick={() => toggleCompleteTask(task.id)}
-                      >
-                        Complete
-                      </button>
-                      <button
-                        className={style.delete}
-                        onClick={() => deleteTask(task.id)}
-                      >
-                        <i className="fa-solid fa-trash"></i>
-                      </button>
-                      <button
-                        className={style.edit}
-                        onClick={() => handleEditForm(task)}
-                      >
-                        <i className="fa-solid fa-pen-to-square"></i>
-                      </button>
-                    </div>
-                  </div>
-                ))
-              )}
+                  ))
+                )}
+              </div>
             </div>
           </div>
         </div>
       </div>
-    </div>
+    </>
   );
 }
